add account to session
