+++ Sources/Network/Models/Session.swift
@@ -10,8 +10,9 @@ import Foundation
 public struct Session: Decodable {
     public let id: String
     public let token: Session.Token
+    public let account: Account
     public let createdAt: Date
-    public let transaction: Session.Transaction?
+    public let transaction: Transaction?
 }
 
 extension Session {
@@ -21,21 +22,5 @@ extension Session {
         public let createdAt: Date
         public let refreshAt: Date
         public let validUntil: Date
-    }
-
-    public struct Transaction: Decodable {
-        public let id: String
-        public let state: State
-        public let amount: Int
-        public let currencyCode: String
-
-        public enum State: String, Decodable {
-            case preauthorizationSuccessful = "PREAUTHORIZATION_SUCCESSFUL"
-            case preauthorizationFailed = "PREAUTHORIZATION_FAILED"
-            case successful = "SUCCESSFUL"
-            case failed = "FAILED"
-            case errored = "ERRORED"
-            case aborted = "ABORTED"
-        }
     }
 }
