+++ Tests/Core/Resources/sessions.json
@@ -9,6 +9,16 @@
           "validUntil": "2022-12-22T09:44:38Z"
       },
       "createdAt": "2022-12-22T09:24:38Z",
+      "account": {
+          "id": "1",
+          "name": "John Doe's Account",
+          "holderName": "John Doe",
+          "currencyCode": "EUR",
+          "createdAt": "2022-12-22T09:24:38Z",
+          "bank": "Bank Name",
+          "iban": "DE123**********",
+          "mandateState": "ACCEPTED"
+      },
       "transaction": {
         "id": "1",
         "state": "PREAUTHORIZATION_SUCCESSFUL",
