+++ Tests/Network/Resources/sessions-post.json
@@ -7,6 +7,16 @@
       "refreshAt": "2022-12-22T09:24:38Z",
       "validUntil": "2022-12-22T09:44:38Z"
   },
+  "account": {
+      "id": "1",
+      "name": "John Doe's Account",
+      "holderName": "John Doe",
+      "currencyCode": "EUR",
+      "createdAt": "2022-12-22T09:24:38Z",
+      "bank": "Bank Name",
+      "iban": "DE123**********",
+      "mandateState": "ACCEPTED"
+  },
   "createdAt": "2022-12-22T09:34:38Z",
   "transaction": null
 }
